+++ server/src/main/java/org/opensearch/search/lookup/SourceLookup.java
@@ -75,8 +75,8 @@ public class SourceLookup implements Map {
         return source;
     }
 
-    public XContentType sourceContentType() {
-        return XContentType.fromMediaType(sourceContentType);
+    public MediaType sourceContentType() {
+        return sourceContentType;
     }
 
     public int docId() {
@@ -155,7 +155,7 @@ public class SourceLookup implements Map {
         this.sourceAsBytes = source;
     }
 
-    public void setSourceContentType(XContentType sourceContentType) {
+    public void setSourceContentType(MediaType sourceContentType) {
         this.sourceContentType = sourceContentType;
     }